+++ src/main/kotlin/lotto/ui/OutputView.kt
@@ -1,7 +1,7 @@
 package lotto.ui
 
-import lotto.Lottoes
-import lotto.Rank
+import lotto.domain.Lottoes
+import lotto.domain.Rank
 
 class OutputView {
     fun printPurchasedQuantity(number: Int) {
@@ -19,9 +19,9 @@ class OutputView {
         println("------------")
         println("3개 일치 (5000원) - ${ranks.count { rank -> rank == Rank.FIFTH }}개")
         println("4개 일치 (50,000원) - ${ranks.count { rank -> rank == Rank.FOURTH }}개")
-        println("5개 일치 (1,500,000원) - ${ranks.count { rank -> rank == lotto.Rank.THIRD }}개")
-        println("5개 일치, 보너스 볼 일치 (30,000,000원) - ${ranks.count { rank -> rank == lotto.Rank.SECOND }}개")
-        println("6개 일치 (2,000,000,000원) - ${ranks.count { rank -> rank == lotto.Rank.FIRST }}개")
+        println("5개 일치 (1,500,000원) - ${ranks.count { rank -> rank == Rank.THIRD }}개")
+        println("5개 일치, 보너스 볼 일치 (30,000,000원) - ${ranks.count { rank -> rank == Rank.SECOND }}개")
+        println("6개 일치 (2,000,000,000원) - ${ranks.count { rank -> rank == Rank.FIRST }}개")
     }
 
     fun printRateOfReturn(money: Int, prizeMoney: Int) {
